locnus/requirements.txt
crispy-bootstrap5
django_extensions
django-tailwind
jupyterlab
jupyter-server<2.0.0
Mastodon.py
pytest-cov
pytest-django
pytest-mock
pytest-randomly
